requirements.txt
django
python-docx
nltk
